+++ pkg/app/util_test.go
@@ -1,0 +1,82 @@
+package app
+
+import (
+	"strconv"
+	"testing"
+
+	"github.com/giantswarm/dex-operator/pkg/key"
+	corev1 "k8s.io/api/core/v1"
+)
+
+func TestGetBaseDomain(t *testing.T) {
+	testCases := []struct {
+		name           string
+		data           map[string]string
+		expectedDomain string
+	}{
+		{
+			name: "case 0",
+			data: map[string]string{
+				key.ValuesConfigMapKey: `
+					something: "12"
+					baseDomain: hello.io
+					somethingelse: "false"
+					object:
+					  yes: no
+					`,
+			},
+			expectedDomain: "hello.io",
+		},
+		{
+			name: "case 1",
+			data: map[string]string{
+				key.ValuesConfigMapKey: `
+					something: "12"
+					somethingelse: "false"
+					object:
+					  yes: no
+					baseDomain: hi.goodday.hello.io
+					`,
+			},
+			expectedDomain: "hi.goodday.hello.io",
+		},
+		{
+			name: "case 2",
+			data: map[string]string{
+				key.ValuesConfigMapKey: `
+					something: "12"
+					somethingelse: "false"
+					object:
+					  yes: no
+					`,
+			},
+			expectedDomain: "",
+		},
+		{
+			name: "case 3",
+			data: map[string]string{
+				key.ValuesConfigMapKey: `
+				baseDomain: hi.goodday.hello.io
+					something: "12"
+					somethingelse: "false"
+					object:
+					  yes: no
+					  baseDomain: no.goodday.hello.io
+					`,
+			},
+			expectedDomain: "hi.goodday.hello.io",
+		},
+	}
+
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			cm := &corev1.ConfigMap{
+				Data: tc.data,
+			}
+			baseDomain := GetBaseDomainFromClusterValues(cm)
+			if baseDomain != tc.expectedDomain {
+				t.Fatalf("Expected %v to be equal to %v", baseDomain, tc.expectedDomain)
+			}
+		})
+	}
+}